Elixir Sudoku v0.1.0 Sudoku.Board

A module containing functions which pertain to working with a Sudoku board.

Functions

clear_cell(board, coords)

Clears the value of the cell at the given coordinates.

Parameters

board: The Sudoku board.

coords: The {x, y} coordinates of the cell that we would like to clear the value of.

Examples

iex(1)> coords = {0, 3}
iex(2)> Sudoku.Board.clear_cell(board, coords)
[[0, 0, 0, 0, 0, 0, 3, 0, 0], ...]

cols_valid?(board)

Returns a boolean indicating if all columns within the given board are valid, which is to say none contain values which repeat.

Parameters

board: The Sudoku board.

Examples

iex> Sudoku.Board.cols_valid?(board)
true

empty?(val)

Returns a boolean indicating if the given value is empty, which is to say that it is equal to 0.

Parameters

val: The value to test.

Examples

iex> board |> get_cell({0, 0}) |> Sudoku.Board.empty?
true

get_board_from_str(str)

Returns a %Sudoku.Board{} struct containing the size of the puzzle and a two-dimensional list of all of its rows generated from the given string representation of the Sudoku puzzle.

Parameters

str: The string representation of the Sudoku puzzle.

Examples

iex(1)> str = "0,0,0,0,9,7,0,0,6,...,9,0,0,6,4,0,0,0,0"
iex(2)> Sudoku.Board.new(str)
%Sudoku.Board{size: 9, rows: [[0, 0, 0, 0, 9, 7, 0, 0, 6], ..., [9, 0, 0, 6, 4, 0, 0, 0, 0]]}

get_cell(board, arg)

Returns the value of the cell at the given coordinates.

Parameters

board: The Sudoku board.

coords: The {x, y} coordinates of the cell that we would like to get the value of.

Examples

iex> Sudoku.Board.get_cell(board, {0, 0})
0

get_col(board, x)

Returns a list of all values within the column at the given index.

Parameters

board: The Sudoku board.

x: The index of the column.

Examples

iex> Sudoku.Board.get_col(board, 0)
[0, 5, 3, 0, 0, 4, 0, 8, 9]

get_cols(board)

Returns a list of all columns within the given board.

Parameters

board: The Sudoku board.

Examples

iex> Sudoku.Board.get_cols(board)
[[0, 5, 3, 0, 0, 4, 0, 8, 9], ...]

get_first_empty_col(board, row_index)

Returns the the index of the first empty column at the within the row at the given index. If there are no empty columns, return nil.

Parameters

board: The Sudoku board.

row_index: The index of the row.

Examples

iex> Sudoku.Board.get_first_empty_col(board, 0)
0

get_legal_values(board, coords)

Returns a list of all legal values for the given {x, y} coordinate.

Parameters

board: The Sudoku board.

coords: The cell we wish to get all legal values for.

Examples

iex> Sudoku.Board.get_legal_values(board, {0, 0})
[1, 2]

get_next_empty_cell(board)

Returns the {x, y} coordinates of the next empty cell in the given board.

Parameters

board: The Sudoku board.

Examples

iex> Sudoku.Board.get_next_empty_cell(board)
{0, 0}

get_next_empty_row(board)

Returns the first row that contains an empty (0) value somewhere. If there are no empty rows, return nil.

Parameters

board: The Sudoku board.

Examples

iex> Sudoku.Board.get_next_empty_row(board)
0

get_row(board, y)

Returns a list of all values within the row at the given index.

Parameters

board: The Sudoku board.

y: The index of the row to get.

Examples

iex> Sudoku.Board.get_row(board, 0)
[0, 0, 0, 0, 9, 7, 0, 0, 6]

get_rows(board)

Returns a list of all rows within the given board.

Parameters

board: The Sudoku board.

Examples

iex> Sudoku.Board.get_rows(board)
[[0, 0, 0, 0, 9, 7, 0, 0, 6], ...]

get_rows_from_str(str)

Returns a two-dimensional list of all rows that exist in the Sudoku puzzle given its string representation.

Parameters

str: The string representation of the Sudoku puzzle.

Examples

iex(1)> str = "0,0,0,0,9,7,0,0,6,...,9,0,0,6,4,0,0,0,0"
iex(2)> Sudoku.Board.get_rows_from_str(str)
[[0, 0, 0, 0, 9, 7, 0, 0, 6], ..., [9, 0, 0, 6, 4, 0, 0, 0, 0]]

get_size(board)

Returns the size of the board.

Parameters

board: The Sudoku board.

Examples

iex> Sudoku.Board.get_size(board)
9

get_size_from_str(str)

Returns the size of the Sudoku board determined from its string representation.

Parameters

str: The string representation of the Sudoku puzzle.

Examples

iex(1)> str = "0,0,0,0,9,7,0,0,6,...,9,0,0,6,4,0,0,0,0"
iex(2)> Sudoku.Board.get_size_from_str(str)
9

get_square(board, arg)

Returns a list of all values within the square which contains the given coordinates.

Parameters

board: The Sudoku board.

coords: The {x, y} coordinates of a cell that exists within the square we would like to get.

Examples

iex> Sudoku.Board.get_square(board, {0, 0})
[0, 0, 0, 5, 0, 0, 3, 0, 0]

get_square_size(board)

Gets the size of a square within the Sudoku board, given its size.

Parameters

board: The Sudoku board.

Examples

iex> Sudoku.Board.get_square_size(9)
3

get_squares(board)

Returns a list of all squares within the given board.

Parameters

board: The Sudoku board.

Examples

iex> Sudoku.Board.get_square(board)
[[0, 0, 0, 5, 0, 0, 3, 0, 0], ...]

get_str_from_board(board)

Transforms the %Sudoku.Board{} struct back into its respective string representation. Good for displaying the solved puzzle as a human-readble string.

Parameters

board: The Sudoku board.

Examples

iex> Sudoku.Board.get_str_from_board(board)
"1,8,4,3,9,7,2,5,6,...,9,3,1,6,4,2,7,8,5"

get_unempty_cells(segment)

Returns a list of all unempty cells within a given segment (row, column, square).

Parameters

segment: A list of values to remove unempty cells from.

Examples

iex(1)> segment = [0, 1, 2, 0, 3, 6, 0, 0, 0]
iex(2)> Sudoku.Board.get_unempty_cells(segment)
[1, 2, 3, 6]

get_unique_cells(segment)

Returns a list of unique values from within a given segment (row, column, square).

Parameters

segment: A list of values to remove unempty cells from.

Examples

iex(1)> segment = [1, 1, 2, 3, 6]
iex(2)> Sudoku.Board.get_unique_cells(segment)
[1, 2, 3, 6]

new(str)

Returns a new Sudoku board strucy generated from the given string representation of the Sudoku puzzle.

Parameters

str: The string representation of the Sudoku puzzle.

Examples

iex(1)> str = "0,0,0,0,9,7,0,0,6,...,9,0,0,6,4,0,0,0,0"
iex(2)> Sudoku.Board.new(str)
%Sudoku.Board{size: 9, rows: [[0, 0, 0, 0, 9, 7, 0, 0, 6], ..., [9, 0, 0, 6, 4, 0, 0, 0, 0]]}

rows_valid?(board)

Returns a boolean indicating if all rows within the given board are valid, which is to say none contain values which repeat.

Parameters

board: The Sudoku board.

Examples

iex> Sudoku.Board.rows_valid?(board)
true

segment_valid?(segment)

Returns a boolean indicating if all values within the given segment (row, column, square) are valid, which is to say it does not contain values which repeat.

Parameters

segment: A list of values to test for uniqueness.

Examples

iex(1)> segment = [0, 1, 2, 0, 3, 6, 0, 0, 0]
iex(2)> Sudoku.Board.segment_valid?(segment)
true

iex(1)> segment = [1, 1, 2, 0, 3, 6, 0, 0, 0]
iex(2)> Sudoku.Board.segment_valid?(segment)
false

segments_valid?(segments)

Returns a boolean indicating if all segments (rows, columns, squares) are valid, which is to say none contain values which repeat.

Parameters

segments: A list of segments to test for uniqueness.

Examples

iex(1)> segments = [[0, 1, 2, 0, 3, 6, 0, 0, 0], ...]
iex(2)> Sudoku.Board.segments_valid?(segments)
true

set_cell(board, arg, val)

Sets the value of the cell at the given coordinates.

Parameters

board: The Sudoku board.

coords: The {x, y} coordinates of the cell that we would like to set the value of.

val: The value to set the cell at the given coordinates to.

Examples

iex(1)> coords = {0, 0}
iex(2)> Sudoku.Board.set_cell(board, coords, 1)
[[1, 0, 0, 5, 0, 0, 3, 0, 0], ...]

squares_valid?(board)

Returns a boolean indicating if all squares within the given board are valid, which is to say none contain values which repeat.

Parameters

board: The Sudoku board.

Examples

iex> Sudoku.Board.squares_valid?(board)
true

valid?(board)

Returns a boolean indicating whether the given board is valid, which is to say all row, column, and square segments do not contain values which repeat.

Parameters

board: The Sudoku board.

Examples

iex> Sudoku.Board.valid?(board)
true