Elixir Sudoku v0.1.0 Sudoku.Solver

A module that solves Sudoku puzzles.

Functions

backtrace(board, list)

Clears the current cell and backtraces up the call tree.

Parameters

board: The Sudoku Board.

history: The solve attempt history.

do_solve(board, history \\ [])

Begins the attempt to solve the Sudoku puzzle by first ensuring that the board is valid. If the board is invalid and contains no history, that is an indicator that the board the user entered is invalid. Otherwise, the current attempted solution is invalid and should be backtraced.

Parameters

board: The Sudoku Board.

history: The solve attempt history.

solve(str)

Attempts to solve the given Sudoku puzzle and if successful will return the string representation of the solved puzzle as read from left ro right, top to bottom.

Parameters

str: The string representation of the Sudoku puzzle.

Examples

iex(1)> str = "0,0,0,0,9,7,0,0,6,...,9,0,0,6,4,0,0,0,0"
iex(2)> Sudoku.Solver.solve(str)
"1,8,4,3,9,7,2,5,6,...,9,3,1,6,4,2,7,8,5"

solve_cell_for_all_values(board, history)

Gets all legal values for the current cell and attempts to solve the puzzle using each value recursively. If no legal values exist, then we must backtrace.

Parameters

board: The Sudoku Board.

history: The solve attempt history.

solve_for_next_legal_value(board, history)

Attempts to solve the puzzle using the next legal value stored in history. If no more legal values exist, then we must backtrace.

Parameters

board: The Sudoku Board.

history: The solve attempt history.

solve_next(board, history)

Gets the coordinates of the next empty cell on the board and attempts to solve it. If there are no remaining empty cells, then congratulations! The puzzle is solved.

Parameters

board: The Sudoku Board.

history: The solve attempt history.